Add files via upload
This is the machine learning modelling to predict Smoker and Non-Smoker using Bio signal data. Models to be compared in this project are Logistic Regression, Random Forest, AdaBoost, and Neural Network. each model use hyperparameter tunning. Base on Recall score, Neural Network is the best model among the other three.